Create program to input name and age
This program prompts the user to input their name and age into the
interactive shell. The program then outputs strings that interact with
each user input string. The first input is the user's name, and program
responds with a greeting. The second input is the user's age, and the
program responds by incrementing the user's age by 1.